requirements.txt
langchain==1.2.6
langchain-openai==1.1.7
python-dotenv==1.1.1
crewai==1.8.1
crewai-tools==1.8.1
ipykernel==7.1.0